pyproject.toml
[tool.poetry]
name = "advocators"
version = "0.1.0"
description = ""
authors = ["1"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
django-jsonfield = "^1.4.1"
Pillow = "^9.2.0"
django-json-field = "^0.5.7"
django-jsonfield-backport = "^1.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
